commit: use %llu and unsigned long long in printf instead of off_t
works better cross architectures

(cherry picked from commit b05ac8826fa0c9b237e99501095b8673050ea81e)

+++ src/commit.c
@@ -391,16 +391,16 @@ all_done:
 		}
 
 		if (apk_verbosity > 1) {
-			apk_message("%s %d packages, %d dirs, %d files, %zu MiB",
+			apk_message("%s %d packages, %d dirs, %d files, %llu MiB",
 				    buf,
 				    installed_packages,
 				    db->installed.stats.dirs,
 				    db->installed.stats.files,
-				    installed_bytes / (1024 * 1024));
+				    (unsigned long long)installed_bytes / (1024 * 1024));
 		} else {
-			apk_message("%s %zu MiB in %d packages",
+			apk_message("%s %llu MiB in %d packages",
 				    buf,
-				    installed_bytes / (1024 * 1024),
+				    (unsigned long long)installed_bytes / (1024 * 1024),
 				    installed_packages);
 		}
 	}
